backend/requirements.txt
# 웹 프레임워크
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1

# 데이터베이스
sqlalchemy==2.0.37
asyncpg==0.30.0
alembic==1.14.1
psycopg2-binary==2.9.10

# 캐시 / 메시지 큐
redis[asyncio]==5.2.1
celery==5.4.0

# AI
anthropic==0.42.0

# 주식 데이터
yfinance>=0.2.50
pandas==2.2.3
numpy==1.26.4

# 기술적 분석
ta==0.11.0

# 뉴스
requests==2.32.3
httpx==0.28.1

# 로깅
loguru==0.7.3

# 환경변수
python-dotenv==1.0.1

# 유틸리티
pytz==2025.1
exchange-calendars>=4.6
